PrizeSystem/Code4Code6/requirements.txt
Flask
flask-sqlalchemy
pymysql
cryptography
mysql-connector-python